+++ src/DotNetWorker.OpenTelemetry/OpenTelemetryConstants.cs
@@ -12,6 +12,10 @@ namespace Microsoft.Azure.Functions.Worker.OpenTelemetry
         internal const string RegionNameEnvVar = "REGION_NAME";
         internal const string ResourceGroupEnvVar = "WEBSITE_RESOURCE_GROUP";
         internal const string OwnerNameEnvVar = "WEBSITE_OWNER_NAME";
+        internal const string ServiceNameEnvVar = "OTEL_SERVICE_NAME";
+        internal const string ResourceAttributeEnvVar = "OTEL_RESOURCE_ATTRIBUTES";
+        internal const string SiteUpdateIdEnvVar = "FUNCTIONS_SITE_UPDATE_ID";
+        internal const string SlotNameEnvVar = "WEBSITE_SLOT_NAME";
         internal const string WorkerDefaultSchemaVersion = "1.37.0";
         internal const string WorkerActivitySourceName = "Microsoft.Azure.Functions.Worker";